AR-722: rename
